Progress percentage and more

+++ levels/backontrack/properties.c
@@ -1,11 +1,12 @@
 #include "soundbank.h"
 // backontrack properties
-const unsigned short backontrack_properties[] = {
+const unsigned int backontrack_properties[] = {
  /*BG color*/      0x5c1f,
  /*GROUND color*/  0x441c,
  /*gamemode*/      0,
  /*speed*/         1,
  /*level height*/  27,
+ /*level width*/   844,
  /*song*/          MOD_BACKOT,
 };
