+++ levels/backontrack/properties.h
@@ -1,5 +1,5 @@
 #pragma once
 
 // backontrack properties
-extern const unsigned short backontrack_properties[];
+extern const unsigned int backontrack_properties[];
 extern const unsigned char backontrack_name[];
