+++ levels/baseafterbase/properties.c
@@ -1,11 +1,12 @@
 #include "soundbank.h"
 // baseafterbase properties
-const unsigned short baseafterbase_properties[] = {
+const unsigned int baseafterbase_properties[] = {
  /*BG color*/      0x7ce5,
  /*GROUND color*/  0x6440,
  /*gamemode*/      0,
  /*speed*/         1,
  /*level height*/  27,
+ /*level width*/   864,
  /*song*/          MOD_STEREOMA,
 };
