+++ levels/baseafterbase/properties.h
@@ -1,5 +1,5 @@
 #pragma once
 
 // baseafterbase properties
-extern const unsigned short baseafterbase_properties[];
+extern const unsigned int baseafterbase_properties[];
 extern const unsigned char baseafterbase_name[];
