+++ levels/clutterfunk/properties.h
@@ -1,5 +1,5 @@
 #pragma once
 
 // clutterfunk properties
-extern const unsigned short clutterfunk_properties[];
+extern const unsigned int clutterfunk_properties[];
 extern const unsigned char clutterfunk_name[];
