+++ levels/cycles/properties.h
@@ -1,5 +1,5 @@
 #pragma once
 
 // cycles properties
-extern const unsigned short cycles_properties[];
+extern const unsigned int cycles_properties[];
 extern const unsigned char cycles_name[];
