+++ levels/dryout/properties.c
@@ -1,11 +1,12 @@
 #include "soundbank.h"
 // dryout properties
-const unsigned short dryout_properties[] = {
+const unsigned int dryout_properties[] = {
  /*BG color*/      0x24bf,
  /*GROUND color*/  0x181f,
  /*gamemode*/      0,
  /*speed*/         1,
  /*level height*/  27,
+ /*level width*/   840,
  /*song*/          MOD_STEREOMA,
 };
