+++ levels/dryout/properties.h
@@ -1,5 +1,5 @@
 #pragma once
 
 // dryout properties
-extern const unsigned short dryout_properties[];
+extern const unsigned int dryout_properties[];
 extern const unsigned char dryout_name[];
