+++ levels/jumper/properties.c
@@ -1,11 +1,12 @@
 #include "soundbank.h"
 // jumper properties
-const unsigned short jumper_properties[] = {
+const unsigned int jumper_properties[] = {
  /*BG color*/      0x7c72,
  /*GROUND color*/  0x640e,
  /*gamemode*/      0,
  /*speed*/         1,
  /*level height*/  27,
+ /*level width*/   892,
  /*song*/          MOD_STEREOMA,
 };
