+++ levels/jumper/properties.h
@@ -1,5 +1,5 @@
 #pragma once
 
 // jumper properties
-extern const unsigned short jumper_properties[];
+extern const unsigned int jumper_properties[];
 extern const unsigned char jumper_name[];
