+++ levels/polargeist/properties.c
@@ -1,11 +1,12 @@
 #include "soundbank.h"
 // polargeist properties
-const unsigned short polargeist_properties[] = {
+const unsigned int polargeist_properties[] = {
  /*BG color*/      0x3e0,
  /*GROUND color*/  0x280,
  /*gamemode*/      0,
  /*speed*/         1,
  /*level height*/  27,
+ /*level width*/   931,
  /*song*/          MOD_STEREOMA,
 };
