+++ levels/polargeist/properties.h
@@ -1,5 +1,5 @@
 #pragma once
 
 // polargeist properties
-extern const unsigned short polargeist_properties[];
+extern const unsigned int polargeist_properties[];
 extern const unsigned char polargeist_name[];
