+++ levels/stereomadness/properties.c
@@ -1,11 +1,12 @@
 #include "soundbank.h"
 // stereomadness properties
-const unsigned short stereomadness_properties[] = {
+const unsigned int stereomadness_properties[] = {
  /*BG color*/      0x7ce5,
  /*GROUND color*/  0x6440,
  /*gamemode*/      0,
  /*speed*/         1,
  /*level height*/  27,
+ /*level width*/   889,
  /*song*/          MOD_STEREOMA,
 };
