+++ levels/stereomadness/properties.h
@@ -1,5 +1,5 @@
 #pragma once
 
 // stereomadness properties
-extern const unsigned short stereomadness_properties[];
+extern const unsigned int stereomadness_properties[];
 extern const unsigned char stereomadness_name[];
