+++ levels/test/properties.c
@@ -1,11 +1,12 @@
 #include "soundbank.h"
 // test properties
-const unsigned short test_properties[] = {
+const unsigned int test_properties[] = {
  /*BG color*/      0x7ce5,
  /*GROUND color*/  0x6440,
  /*gamemode*/      0,
  /*speed*/         1,
  /*level height*/  27,
+ /*level width*/   895,
  /*song*/          MOD_STEREOMA,
 };
