+++ levels/test/properties.h
@@ -1,5 +1,5 @@
 #pragma once
 
 // test properties
-extern const unsigned short test_properties[];
+extern const unsigned int test_properties[];
 extern const unsigned char test_name[];
