+++ levels/timemachine/properties.c
@@ -1,11 +1,12 @@
 #include "soundbank.h"
 // timemachine properties
-const unsigned short timemachine_properties[] = {
+const unsigned int timemachine_properties[] = {
  /*BG color*/      0x3ce,
  /*GROUND color*/  0x246,
  /*gamemode*/      0,
  /*speed*/         1,
  /*level height*/  27,
+ /*level width*/   993,
  /*song*/          MOD_STEREOMA,
 };
