+++ levels/timemachine/properties.h
@@ -1,5 +1,5 @@
 #pragma once
 
 // timemachine properties
-extern const unsigned short timemachine_properties[];
+extern const unsigned int timemachine_properties[];
 extern const unsigned char timemachine_name[];
