+++ levels/xstep/properties.c
@@ -1,11 +1,12 @@
 #include "soundbank.h"
 // xstep properties
-const unsigned short xstep_properties[] = {
+const unsigned int xstep_properties[] = {
  /*BG color*/      0x4600,
  /*GROUND color*/  0x2100,
  /*gamemode*/      0,
  /*speed*/         1,
  /*level height*/  27,
+ /*level width*/   841,
  /*song*/          MOD_XSTEP,
 };
